requirements.txt
streamlit
streamlit-option-menu
pandasai
pandasai[connectors]
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
mysql-connector-python
pymysql